+++ Alicargo.Database/Tables/Event/EventEmailRecipient.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[EventEmailRecipient] (
 	[RoleId] INT NOT NULL,
-	[EventTypeId] BIGINT NOT NULL,
+	[EventTypeId] INT NOT NULL,
 
 	CONSTRAINT [PK_dbo.EventEmailRecipient] PRIMARY KEY CLUSTERED ([RoleId] ASC, [EventTypeId] ASC),
 );
